Integrate on an step only.
