+++ src/test/resources/smartHouseFeatures/AccidentCase3.feature
@@ -2,7 +2,9 @@
 Feature: SHAccident
 
   Scenario: The user close the dialog to register accidents from header 'x' button and 'Cancelar' button
-    Given 'SMARTHOUSE' page accidents is loaded
+    Given Given I loging to 'SMARTHOUSE' page
+    And I fill properly credentials with admin user
+    And I select accident menu 'Accidentes'
     And I click on 'Registrar accidente' button
     And Close modal 'Registro de Accidente' from 'x' button
     And I click on 'Registrar accidente' button
